dataforge/requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.8.1
apache-airflow-providers-postgres==5.10.0
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
yfinance==0.2.32
great-expectations==0.17.12
pytest==7.4.3
pytest-cov==4.1.0
structlog==23.2.0
setuptools==69.0.2
minio==7.2.0
duckdb==0.9.2
sqlalchemy==2.0.23
